Extract Abstract Doctrine Middleware
